Renaming bookend/share into amp-story-bookend/share.
